requirements.txt
streamlit==1.32.2
together==0.2.10
python-dotenv==1.0.0